Recreated the Footer
In the head section of the index page, I added the FontAwesome link to include social media icons and a Google Font link for the font used in the footer. Following that, I implemented the footer code and included two images: indianFlag.png and logo-protrack.png for use in the footer.